LPS-89224 Inline keyTransformer class
